Move get-html into core/dom
